home page done

+++ app/Http/Controllers/V1/Admin/HomePageController.php
@@ -5,8 +5,8 @@ namespace App\Http\Controllers\V1\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\HomeAbout;
 use App\Models\ImageVideo;
-use App\Models\Slider;
 use App\Models\LatestPodcast;
+use App\Models\Slider;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
@@ -34,9 +34,9 @@ class HomePageController extends Controller
     {
         $request->validate([
 
-            'content_en' => 'nullable|min:3',
-            'content_hi' => 'nullable|min:3',
-            'content_gu' => 'nullable|min:3',
+            'content_en' => 'required|min:3',
+            'content_hi' => 'required|min:3',
+            'content_gu' => 'required|min:3',
             'image' => 'required|mimes:png,jpg,jpeg||max:2048',
 
         ]);
@@ -67,10 +67,10 @@ class HomePageController extends Controller
 
         $request->validate([
 
-            'content_en' => 'nullable|min:3',
-            'content_hi' => 'nullable|min:3',
-            'content_gu' => 'nullable|min:3',
-            'image' => 'required|mimes:png,jpg,jpeg||max:2048',
+            'content_en' => 'required|min:3',
+            'content_hi' => 'required|min:3',
+            'content_gu' => 'required|min:3',
+            'image' => 'nullable|mimes:png,jpg,jpeg||max:2048',
 
         ]);
         try {
@@ -110,8 +110,6 @@ class HomePageController extends Controller
         $data = HomeAbout::all();
         return view('Admin.Home.homeAbout', compact('data'));
     }
-
-    
 
     public function homeAbout(Request $request)
     {
@@ -184,12 +182,14 @@ class HomePageController extends Controller
 
         ]);
         try {
-           
-            $media= new ImageVideo();
-            $imageName = time() . '.' . $request->image->extension();
-            $request->image->move(public_path('Home_Images'), $imageName);
-            $media->image=$imageName;
-            $media->video_link=$request->video_link;
+
+            $media = new ImageVideo();
+            if ($request->hasFile('image')) {
+                $imageName = time() . '.' . $request->image->extension();
+                $request->image->move(public_path('Home_Images'), $imageName);
+                $media->image = $imageName;
+            }
+            $media->video_link = $request->video_link;
             $media->save();
 
             return redirect()->route('filmvideo.index')->with('message', 'Add Successfully..');
@@ -199,31 +199,34 @@ class HomePageController extends Controller
     }
 
     public function editImageVideo($id)
-    {   
-        $data=ImageVideo::find($id);
-        return view('Admin.Home.edit_imagevideo',compact('data'));
+    {
+        $data = ImageVideo::find($id);
+        return view('Admin.Home.edit_imagevideo', compact('data'));
 
     }
-    public function updateImageVideo(Request $request,$id)
+    public function updateImageVideo(Request $request, $id)
     {
         $request->validate([
 
             "video_link" => "required|url",
-            'image' => 'required|mimes:png,jpg,jpeg||max:2048',
+            'image' => 'nullable|mimes:png,jpg,jpeg||max:2048',
 
         ]);
         try {
-           
-            $data=  ImageVideo::find($id);
-            $path = public_path('Home_Images/' . $data->image);
 
-        if (File::exists($path)) {
-            unlink($path);
-        }
-            $imageName = time() . '.' . $request->image->extension();
-            $request->image->move(public_path('Home_Images'), $imageName);
-            $data->image=$imageName;
-            $data->video_link=$request->video_link;
+            $data = ImageVideo::find($id);
+            if ($request->hasFile('image')) {
+                $path = public_path('Home_Images/' . $data->image);
+
+                if (File::exists($path)) {
+                    unlink($path);
+                }
+
+                $imageName = time() . '.' . $request->image->extension();
+                $request->image->move(public_path('Home_Images'), $imageName);
+                $data->image = $imageName;
+            }
+            $data->video_link = $request->video_link;
             $data->save();
 
             return redirect()->route('filmvideo.index')->with('message', 'update Successfully..');
@@ -231,9 +234,9 @@ class HomePageController extends Controller
             return $e->getMessage();
         }
     }
-    public  function deleteImageVideo($id)
+    public function deleteImageVideo($id)
     {
-        $data=ImageVideo::find($id);
+        $data = ImageVideo::find($id);
         $path = public_path('Home_Images/' . $data->image);
 
         if (File::exists($path)) {
@@ -243,15 +246,14 @@ class HomePageController extends Controller
         return redirect()->route('filmvideo.index')->with('message', 'Delete Successfully..');
     }
 
-
     public function indexProdcast()
     {
-        $data=LatestPodcast::all();
-        return view('Admin.Home.index_prodcast',compact('data'));
+        $data = LatestPodcast::whereNot('id', [1])->get();
+        return view('Admin.Home.index_prodcast', compact('data'));
     }
     public function addProdcast()
     {
-       
+
         return view('Admin.Home.add_prodcast');
     }
     public function storeProdcast(Request $request)
@@ -260,30 +262,144 @@ class HomePageController extends Controller
 
             "music" => "required|mimes:mp3",
             'image' => 'required|mimes:png,jpg,jpeg||max:2048',
-            'content_en'=>'required|min:3',
-            'content_hi'=>'required|min:3',
-            'content_gu'=>'required|min:3',
+            'content_en' => 'required|min:3',
+            'content_hi' => 'required|min:3',
+            'content_gu' => 'required|min:3',
+            'date' => 'required',
 
         ]);
         try {
-           
-            $data= new ImageVideo();
-            $imageName = time() . '.' . $request->image->extension();
-            $request->image->move(public_path('Home_Images'), $imageName);
-            $data->image=$imageName;
-            $data->music=$request->music;
-            $data->content_en=$request->content_en;
-            $data->content_hi=$request->content_hi;
-            $data->content_gu=$request->content_gu;
 
+            $data = new LatestPodcast();
 
+            if ($request->hasFile('image')) {
+                $imageName = time() . '.' . $request->image->extension();
+                $request->image->move(public_path('prodcast_image'), $imageName);
+                $data->image = $imageName;
+            }
+
+            if ($request->hasFile('music')) {
+                $music = time() . '.' . $request->music->extension();
+                $request->music->move(public_path('latestProdcast_Audio'), $music);
+                $data->music = $request->music;
+            }
+            $data->content_en = $request->content_en;
+            $data->content_hi = $request->content_hi;
+            $data->content_gu = $request->content_gu;
+            $data->date = $request->date;
             $data->save();
 
             return redirect()->route('prodcast.index')->with('message', 'Add Successfully..');
         } catch (Exception $e) {
             return $e->getMessage();
         }
-       
+
     }
-      
+
+    public function prodcastImage()
+    {
+
+        return view('Admin.Home.prodcast_image');
+    }
+    public function prodcastImageManage(Request $request)
+    {
+        $request->validate([
+
+            'image' => 'required|mimes:png,jpg,jpeg||max:2048',
+
+        ]);
+        try {
+            $imageGet = LatestPodcast::find(1);
+
+            if ($request->hasFile('image')) {
+                $path = public_path('prodcast_image/' . $imageGet->image);
+                if (File::exists($path)) {
+                    unlink($path);
+                }
+            }
+
+            $imageName = time() . '.' . $request->image->extension();
+            $request->image->move(public_path('prodcast_image'), $imageName);
+
+            $data = LatestPodcast::updateOrCreate(
+
+                ['id' => '1'],
+                [
+
+                    'image' => $imageName,
+
+                ]
+            );
+
+            return redirect()->route('prodcast.index')->with('message', ' Successfully..');
+
+        } catch (Exception $e) {
+            return $e->getMessage();
+        }
+    }
+
+    public function editLatestProdcast($id)
+    {
+        $data = LatestPodcast::find($id);
+        return view('Admin.Home.edit_latestprodcast', compact('data'));
+    }
+    public function updateLatestProdcast(Request $request, $id)
+    {
+
+        $request->validate([
+
+            "music" => "nullable|mimes:mp3",
+            'image' => 'nullable|mimes:png,jpg,jpeg||max:2048',
+            'content_en' => 'required|min:3',
+            'content_hi' => 'required|min:3',
+            'content_gu' => 'required|min:3',
+            'date' => 'required',
+
+        ]);
+        try {
+
+            $data = LatestPodcast::find($id);
+
+            if ($request->hasFile('image')) {
+
+                $path = public_path('prodcast_image/' . $data->image);
+                if (File::exists($path)) {
+                    unlink($path);
+                }
+                $imageName = time() . '.' . $request->image->extension();
+                $request->image->move(public_path('prodcast_image'), $imageName);
+                $data->image = $imageName;
+            }
+
+            if ($request->hasFile('music')) {
+                $path = public_path('latestProdcast_Audio/' . $data->music);
+                if (File::exists($path)) {
+                    unlink($path);
+                }
+                $music = time() . '.' . $request->music->extension();
+                $request->music->move(public_path('latestProdcast_Audio'), $music);
+                $data->music = $request->music;
+            }
+            $data->content_en = $request->content_en;
+            $data->content_hi = $request->content_hi;
+            $data->content_gu = $request->content_gu;
+            $data->date = $request->date;
+            $data->save();
+
+            return redirect()->route('prodcast.index')->with('message', 'Update Successfully..');
+        } catch (Exception $e) {
+            return $e->getMessage();
+        }
+
+    }
+
+    public function deleteLatestProdcast($id)
+    {   
+        
+        $data = LatestPodcast::find($id);
+        $data->delete();
+        return redirect()->route('prodcast.index')->with('message', ' Successfully Delete..');
+
+
+    }
 }
